refactor: Rename exceptions dir
refactor: Rename exceptions dir

refactor: Rename exceptions dir

Can be squashed into previous commits
