+++ lib/src/digit_span_task/components/exceptions/errors.dart
