+++ lib/src/digit_span_task/components/stim/stim_controller.dart
@@ -1,7 +1,7 @@
 import 'package:digit_span_tasks/src/digit_span_task/components/config/session_type.dart';
 import 'package:get/get.dart';
 import 'package:digit_span_tasks/src/digit_span_task/components/activity/activity_controller.dart';
-import 'package:digit_span_tasks/src/digit_span_task/components/errors/errors.dart';
+import 'package:digit_span_tasks/src/digit_span_task/components/exceptions/errors.dart';
 import 'package:digit_span_tasks/src/digit_span_task/components/activity/task_step.dart';
 import 'package:stimuli/errors.dart';
 import 'package:stimuli/stimuli.dart';
